Update UserController to use AppDbContext in CreateUser
This commit modifies the `UserController.cs` file by adding a `using` directive for `TuneMates_Backend.DataBase`. The `CreateUser` method has been updated to accept an `AppDbContext db` parameter. Inside this method, a new `User` object is created with predefined username and email, added to the database, and changes are saved. The method now returns a `TypedResults.Ok()` result.