Remove user management routes and related functions
This commit removes two API routes and their associated asynchronous functions, TestFunction and Test2Function, from Program.cs. TestFunction was responsible for creating a new user in the database, while Test2Function verified if a provided password matched a stored password. These changes suggest a redesign or simplification of the user management logic in the application.