Update user management routes with authorization
Modified user management methods to use more general routes that require authorization. Removed old routes for editing a user, changing a password, and deleting a user. Added new routes for editing a user, changing a password, and deleting a user without specifying an ID, all of which now require authorization.