Update user management routes to use "/me" prefix
This commit modifies the user management routes. The previous routes for editing a user, changing the password, and deleting a user have been replaced with new routes that use the "/me" prefix. This means that user edit, password change, and delete operations are now performed on the "/me" path instead of the root "/".