Add AddSongToPlaylistAsync method to SpotifyApi.cs
This commit introduces an asynchronous method `AddSongToPlaylistAsync` in the `TuneMates_Backend.Utils` namespace of the `SpotifyApi.cs` file. The method allows adding a song to a Spotify playlist by first checking that the playlist and song IDs are not empty. It constructs the song URI if necessary, retrieves an access token, creates an HTTP request to add the song to the playlist, and returns a boolean indicating whether the operation was successful or not.