Add API route for adding songs to Spotify playlist
This commit introduces a new API route to add a song directly to a room's Spotify playlist using the song ID from the database. The song will be marked as "Approved" after being added to the playlist. Additionally, the documentation has been updated to specify that there must be a space between `Bearer` and the token.