Refactor SpotifyApi instantiation in SongController
This commit removes the use of HttpClient in the SongController
when creating an instance of SpotifyApi. Now, SpotifyApi is
instantiated solely with the database context (db) and
configuration (cfg), simplifying the code and improving resource
management and performance.

Additionally, the constructor of SpotifyApi has been modified
to no longer accept HttpClient as a parameter. Instead, a new
instance of HttpClient is created within the constructor,
encapsulating its management within SpotifyApi and facilitating
testing and usage without requiring external configuration.