Add IMemoryCache support in SongController methods
This update introduces a dependency on IMemoryCache in several methods of the SongController.cs file.
This allows for memory caching to optimize calls to the Spotify API. The method signatures for AddSongToPlaylist,
AddSongToRoom, and SearchSongs have been updated to include an IMemoryCache cache parameter, replacing the
IConfiguration cfg parameter in the instantiation of SpotifyApi. This change indicates that memory caching
will be used to temporarily store data, potentially improving performance by reducing the number of calls
to the external API.