Add SpotifyId property to UserResponse class
A new property `SpotifyId` of type `string` was added to the `UserResponse` class, initialized with a default value of `string.Empty`. The constructor of the class was updated to map the `SpotifyId` property from the `user` object. This change enables support for storing and managing a Spotify identifier for users, facilitating potential integrations or Spotify-related features.