Fetch Spotify user ID during token retrieval
Added functionality to fetch the Spotify user ID by making a GET request to the `https://api.spotify.com/v1/me` endpoint using the access token. The user ID is extracted from the JSON response and assigned to the `SpotifyUserId` property of the token. Errors during this process are gracefully handled to avoid interrupting the flow.